irkScan

+++ frameworks/inner/ipc/parcel/bluetooth_raw_address.cpp
@@ -23,7 +23,7 @@ bool BluetoothRawAddress::Marshalling(Parcel &parcel) const
     if (!parcel.WriteString(address_)) {
         return false;
     }
-    if (!parcel.WriteInt32(addressType_)) {
+    if (!parcel.WriteUint8(addressType_)) {
         return false;
     }
     return true;
@@ -49,7 +49,7 @@ bool BluetoothRawAddress::ReadFromParcel(Parcel &parcel)
     if (!parcel.ReadString(address_)) {
         return false;
     }
-    if (!parcel.ReadInt32(addressType_)) {
+    if (!parcel.ReadUint8(addressType_)) {
         return false;
     }
     return true;
